refactor(server): proper Python package with tool-calling, tests
- src/dflash/server/ replaces scripts/server.py + scripts/server_tools.py
- server/parsing.py: Qwen3.x tool-call + reasoning parsers (pure fns, no deps)
- server/schemas.py: pydantic models for OpenAI + Anthropic APIs
- Entry point: dflash-server = "dflash.server:main" (uv run dflash-server)
- tests/: 33 tests covering parsing and all HTTP endpoints
- Consolidates tool calling, streaming state machine, GGUF model name,
  port 1236, 128K ctx defaults from server_tools.py into single package
